+++ anpe/__init__.py
@@ -9,7 +9,7 @@ Accurate noun phrase extraction using the Berkeley Neural Parser.
 import logging
 
 # Define package version
-__version__ = "1.1.2"
+__version__ = "1.1.3"
 
 # Import key classes/functions for easier access
 from anpe.extractor import ANPEExtractor, extract, export